+++ app/src/main/java/com/example/mwt/fragments/drinkingstatistics/DrinkingStatisticsFragment.kt
@@ -48,7 +48,7 @@ class DrinkingStatisticsFragment : Fragment() {
             val data = it.reversed()
             if(data.isNotEmpty()){
                 for (i in data.indices) {
-                    barEntries.add(BarEntry(i.toFloat(), data[i].progress))
+                    barEntries.add(BarEntry(i.toFloat(), data[i].progress.toFloat()))
                     barDate.add(data[i].date)
                 }
             }else{
